termassignment/backend/lambda_textract/requirements.txt
boto3
pprintpp
pillow
